wmediumd: fix unintentional timerfd busy loop
When a timerfd event is raised, the underlying object contains
a u64 that counts the number of expirations since the last
timerfd_settime().  If we don't read that value, it stays readable
and the subsequent poll will return immediately, causing us to
just burn CPU instead of waiting for the next expiration.  Fix by
reading the fd here.