wmediumd: track dynamic station addresses
With newer kernels, mac80211_hwsim indicates to us when it starts
and stops using a certain MAC address on a certain radio. Without
that, there's a need to configure each MAC address as a station,
which is not only incorrect (a station can have multiple addresses
and wouldn't have a different position or different frame loss
probabilities between them) but also impractical, e.g. for random
scanning.

Add the necessary code to track and use this information, and look
up sender station by the HW address and destination by any of the
used addresses.

Still keep looking up by source address as well to not break any
existing use cases.