wmediumd: properly wait for control socket ACK
When the control socket is used in conjunction with a time
controller, then it may happen that we have

wmediumd               app                  time controller
  ------ message ------>
  (wait for ack)
                         ----- request runtime --->
  <-------------- update scheduler ----------------
                                              (wait for ack)

and deadlock here, because the update scheduler message
isn't processed by wmediumd.

Fix this by properly deferring to the mainloop when waiting
for an ACK message on the control socket.